Fix throwing the wrong error when an S0_Legacy key is missing.
If *no* securityKeys whatsoever have been provided in a config file,
e.g., if securityKeys is null, then the current code will throw a
null dereferencing error instead of the intended "S0_Legacy key is
missing" error.  A '?' in the right spot should take care of this.

Also, remove the "!options.networkKey" clause, which is redundant
(because this is the ELSE to "if (options.networkKey) ...").